move bcrypt hashing package to subfolder
