+++ hashing/bcrypt/bcrypt.go
@@ -1,6 +1,8 @@
-package hashing
+package bcrypt
 
-import "golang.org/x/crypto/bcrypt"
+import (
+	"golang.org/x/crypto/bcrypt"
+)
 
 type bcryptHasher struct {
 }
@@ -13,7 +15,7 @@ func (h *bcryptHasher) Verify(hash string, text string) error {
 	return bcrypt.CompareHashAndPassword([]byte(hash), []byte(text))
 }
 
-func NewBcryptHasher() *bcryptHasher {
+func NewHasher() *bcryptHasher {
 	h := &bcryptHasher{}
 	return h
 }
